requirements.txt
appdirs==1.4.3
blinker==1.4
blist==1.3.6
cassandra-driver==3.8.1
click==6.7
Flask==0.12.1
Flask-CQLAlchemy==1.2.0
Flask-Login==0.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
futures==3.0.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
Whoosh==2.7.4
WTForms==2.1
